added rough tag function
